Issue1019-test-mixed-dim-gravity (#1029)
* MAINT: Refactor extract compute geometry from set projections for BoundaryGrids

* BUG: Switch from avg to int for mortar normal projection

* BUG: Two vector source mistakes

* TST: Move and rework tests for gravity

* TST: Temporarly rm benchmark test

* BUG: Dim of bound vec source for 0d grids

* Update src/porepy/grids/boundary_grid.py


* Update src/porepy/models/geometry.py

* Apply suggestions from code review

* STY: flake8 and black

* BUG: Correct type comparison.

---------

+++ tests/fracs/test_wells.py
@@ -228,4 +228,4 @@ def test_add_one_well_with_matrix(get_mdg) -> None:
     assert len(well_boundaries) == 1
     for well_bg in well_boundaries:
         # num_cells is one of the attributes that are initialized lazily.
-        assert isinstance(well_bg.num_cells, int)
+        assert isinstance(well_bg.num_cells, np.int64)
